requirements.txt
fastapi
uvicorn
faststream[rabbit]
docker 
requests
pyyaml
pyppeteer